pyproject.toml
[project]
name = "ai-engine"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.116.1",
    "uvicorn>=0.35.0",
    "openai>=1.3.0",
    "pydantic>=2.5.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.25.0",
    "tavily-python==0.3.3",
    "langgraph>=0.2.0",
]
